--- assets/collektive.css
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2024, Danilo Pianini, Nicolas Farabegoli, Elisa Tronetti,
- * and all authors listed in the `build.gradle.kts` and the generated `pom.xml` file.
- *
- * This file is part of Collektive, and is distributed under the terms of the Apache License 2.0,
- * as described in the LICENSE file in this project's repository's top directory.
- */
-
-:root {
-    --dokka-logo-image-url: url('../images/logo.svg');
-    --dokka-logo-height: 50px;
-    --dokka-logo-width: 50px;
-}
