+++ compiler-plugin/src/main/kotlin/it/unibo/collektive/utils/common/CollektiveIrUtils.kt
@@ -1,6 +1,7 @@
 package it.unibo.collektive.utils.common
 
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
+import org.jetbrains.kotlin.backend.jvm.ir.receiverAndArgs
 import org.jetbrains.kotlin.ir.IrElement
 import org.jetbrains.kotlin.ir.backend.js.utils.typeArguments
 import org.jetbrains.kotlin.ir.builders.IrBlockBodyBuilder
@@ -13,7 +14,6 @@ import org.jetbrains.kotlin.ir.types.IrType
 import org.jetbrains.kotlin.ir.types.classFqName
 import org.jetbrains.kotlin.ir.types.classifierOrNull
 import org.jetbrains.kotlin.ir.util.kotlinFqName
-import org.jetbrains.kotlin.ir.util.receiverAndArgs
 
 internal fun IrType.isAssignableFrom(other: IrType): Boolean =
     classifierOrNull?.let { base ->
