+++ gradle.properties
@@ -1,6 +1,5 @@
 android.builder.sdkDownload=true
 org.gradle.jvmargs=-XX:MaxMetaspaceSize=2g -Xmx2g
-org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
 # kotlin
 kotlin.code.style=official
 kotlin.native.ignoreDisabledTargets=true
